Add files via upload
Stopwatch
A stopwatch web application uses HTML, CSS, and JavaScript. HTML to structure the elements of the application. By implementing functions for starting, pausing, and resetting the stopwatch, as well as tracking and displaying lap times, users can accurately measure and record time intervals. With these technologies and functionalities, it is built into an interactive and user-friendly stopwatch web application.







web Application